All functions are now calculated using doubles. Main Menu is customizable. Better performance and view

+++ core/src/ru/maklas/genetics/utils/functions/GraphFunction.java
@@ -2,6 +2,6 @@ package ru.maklas.genetics.utils.functions;
 
 public interface GraphFunction {
 
-    float f(float x);
+    double f(double x);
 
 }
